Data.BigWord

Description

Fixed size bit vectors using type-level naturals.

The W newtype

data W n

Type representing a sequence of n bits, or a non-negative integer smaller than 2^n.

Operations

(>+<) :: forall n m o. (KnownNat m, KnownNat n, KnownNat o, o ~ (m + n), o ~ (n + m)) => W m -> W n -> W o

Appends two W's, treating the second's bits as more significant.

Example usage:

   import Network.Socket

   fromHostAddress6 :: HostAddress6 -> W 128
   fromHostAddress6 (a, b, c, d) = f a >+< f b >+< f c >+< f d
     where
       f = fromIntegral :: Word32 -> W 32

split :: forall n m o. (KnownNat m, KnownNat n, KnownNat o, o ~ (m + n), o ~ (n + m)) => W o -> (W m, W n)

The inverse of >+<

   forall a b. split (a >+< b) == (a, b)

Example usage:

   import Network.Socket

   toHostAddress6 :: W 128 -> HostAddress6
   toHostAddress6 w =  (f a, f b, f c, f d)
     where
       f = fromIntegral :: W 32 -> Word32
       (a, x) = split w
       (b, y) = split x
       (c, d) = split y

Utility functions

assembleL :: (Applicative f, d :|: n) => f (W d) -> f (W n)

Transforms an applicative action that results in a W d to on that results in a W n, provided that d|n (hence the :|: constraint), treating the first results as less significant.

Example using attoparsec to parse a little-endian unsigned 128-bit integer:

   import Data.Attoparsec.ByteString
   import Data.Word

   anyWord128LE :: Parser (W 128)
   anyWord128LE = assembleL $ fmap (fromIntegral :: Word8 -> W 8) anyWord8

assembleR :: (Applicative f, d :|: n) => f (W d) -> f (W n)

Same as assembleL, but treats the first results a more significant>

Here's the example above, modified to parse in network-byte order:

   anyWord128BE :: Parser (W 128)
   anyWord128B  = assembleR $ fmap (fromIntegral :: Word8 -> W 8) anyWord8

disassembleL :: (Monoid m, d :|: n) => (W d -> m) -> W n -> m

Breaks a W n into its constituent d-sized chunks, and combines them according to the provided monoid.

Example using a bytestrings

   import Data.ByteString.Builder

   word128LE :: W 128 -> Builder
   word128LE = disassembleL (word8 . (fromIntegral :: W 8 -> Word8))

disassembleR :: (Monoid m, d :|: n) => (W d -> m) -> W n -> m

Same as disassembleL, but mappends from right to left.

disassembleL's example adjusted to build in network-byte order:

   import Data.ByteString.Builder

   word128BE :: W 128 -> Builder
   word128BE = disassembleR (word8 . (fromIntegral :: W 8 -> Word8))