Data.BigWord

Description

Fixed-size bit vectors using type-level naturals.

Documentation

data BigWord n

Type representing a sequence of n bits, or a non-negative integer smaller than 2^n

(>+<) :: forall n m. (KnownNat n, KnownNat m, KnownNat (n + m)) => BigWord n -> BigWord m -> BigWord (n + m)

Appends two BigWords, treating the second's bits as more significant.

takeBE :: (Applicative f, KnownNat n) => f Word8 -> f (BigWord n)

Transforms an action that results in an octet to one that results in a BigWord of the specified size. Octets resulting from the input action occur in network-byte order. If n is not a multiple of 8, leftover bits are truncated.

takeLE :: (Applicative f, KnownNat n) => f Word8 -> f (BigWord n)

Transforms an action that results in an octet to one that results in a BigWord of the specified size. Octets resulting from the input action occur in little-endian order. If n is not a multiple of 8, leftover bits are truncated.

giveBE :: forall m n. (Monoid m, KnownNat n) => (Word8 -> m) -> BigWord n -> m

Break a BigWord into its constituent octets, and combine then as a monoid in network-byte order. If n is not a multiple of 8, missing bits are treated as 0.

giveLE :: (Monoid m, KnownNat n) => (Word8 -> m) -> BigWord n -> m

Break a BigWord into its constituent octets, and combine then as a monoid in little-endian order. If n is not a multiple of 8, missing bits are treated as 0.