Data.BigWord

Description

Fixed size bit vectors using type-level naturals.

The W newtype

data W n

Type representing a sequence of n bits, or a non-negative integer smaller than 2^n.

Operations

(>+<) :: forall n m. (KnownNat m, KnownNat n, KnownNat (m + n)) => W m -> W n -> W (m + n)

Appends two W's, treating the second's bits as more significant.

Example usage:

   import Network.Socket

   fromHostAddress6 :: HostAddress6 -> W 128
   fromHostAddress6 (a, b, c, d) = f a >+< f b >+< f c >+< f d
     where
       f = fromIntegral :: Word32 -> W 32

split :: forall n m. (KnownNat m, KnownNat n, KnownNat (m + n)) => W (m + n) -> (W m, W n)

The inverse of >+<

   forall a b. split (a >+< b) == (a, b)

Example usage:

   import Network.Socket

   toHostAddress6 :: W 128 -> HostAddress6
   toHostAddress6 w =  (f a, f b, f c, f d)
     where
       f = fromIntegral :: W 32 -> Word32
       (a, x) = split w
       (b, y) = split x
       (c, d) = split y

Utility functions

accumulate :: (Applicative f, KnownNat m, KnownNat n) => f (W n) -> f (W m)

Transforms an applicative action that results in a W n to one that results in a W m.

W n's are accumulated with significance increasing from left to right.

If n does not divide m, leftover bits are truncated.

Example usage, parsing a little endian 160-bit word:

   import Data.Attoparsec.ByteString

   parseW160 :: Parser (W 610)
   parseW160 = accumulate $ (fromIntegral :: Word8 -> W 8) <$> anyWord8

accumulate' :: (Applicative f, KnownNat m, KnownNat n) => f (W n) -> f (W m)

Same as accumulate, but gathers in the opposite order

chunks :: forall a m n. (Monoid a, KnownNat m, KnownNat n) => (W n -> a) -> W m -> a

Break a W m into its constituent W n's, and combine using the supplied monoid

Chunks are appended with significance increasing from left to right.

If n does not divide m, missing bits are set to 0.

Example usage, building a little endian 160-bit word:

   import Data.ByteString.Builder

   buildW160 :: W 160 -> Builder
   buildW160 = chunks $ word8 . (fromIntegral :: W 8 -> Word8)

chunks' :: forall a m n. (Monoid a, KnownNat m, KnownNat n) => (W n -> a) -> W m -> a

Same as chunks, but appends in opposite order